requirements-oauth.txt
# Core UAP dependencies
click>=8.0.0
rich>=13.0.0
pyyaml>=6.0
pydantic>=2.0.0

# Google OAuth
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Gemini AI SDK
google-generativeai>=0.3.0

# Web dashboard
streamlit>=1.30.0
requests>=2.28.0

# LLM backends (optional)
groq>=0.4.0
